refactor(app): slim AppDelegate by extracting lifecycle services
AppDelegate had grown to ~200 lines covering appearance observation,
debug menu + log export, Sparkle bootstrap, helper version checks,
and a runtime toolbar fix. Each responsibility now lives in its own
@MainActor controller under App/ (AppearanceController,
DebugMenuController, HelperServiceVersionChecker, the standalone
NSToolbarItemViewerOverflowFix), registered via @Dependency.

AppDelegate reduces to a thin dispatch shell: lifecycle hooks +
one-line start/install/check calls on injected services + IBActions.
